add parameter route middleware for 'atm_id'
This commit adds a parameter route middleware for any url that contains
':atm-id'.  The middleware does some simple checking to ensure that the
requested id exists.  This also modifies the way the server is started
by adding a bin/www.js file.  This is modeled after the generated
express framework.