add ATM controller and util/
This commit adds an ATM controller file to abstract the logic
in the atm routes file.  Also creates a util directory and moves the
datamanager there.  The datamanager was also modified to utilize the
'readFileSync' method of the 'fs' module vice the async version.  This
makes initializing the data from a .json file a bit simplier.