add test for API DELETE verb
This commit adds a test for the DELETE verb for the API.  This deletes
an atm transaction from the array and then does a GET request to verify
the length of the array is reduced by one.  Also adds a space after the
function keyword but before the left paren to make the code style follow
Crockford's conventions.