fix test to point to correct url
This commit fixes an oversight in the server spec file that was
doing a GET request to the wrong url--not testing the intended
route.  Also modifies the 'param' middleware functions to forward
on an err obj (which will be caught by the error handles) vice
returning a simple json obj with a msg.