configure project with 'test' env for data files
At this point in development, there was not a clear distinction
between the data json files for the test environment and the
development environment.  This commit makes that distinction by
setting the process.env.NODE_ENV environment variable to 'test'
when the test is being run.  Based on the value of NODE_ENV,
the appropriate inputFile and outputFile is assigned.

* establishes a 'test' NODE_ENV when testing

* adds a data configuration file

* assigns the appropriate inputFile and outputFile (json) based
  on the value of NODE_ENV

Note: may want to delete the outputFile that is generated by the
test suite.