add ability to update 'totalSpent' property
Added a 'totalSpent' property to each ATM object and a method to update
that property.  For the data that is initially loaded in from the
external json file, it needed to have access to that method to update
its 'totalSpent' property.  To do this, we needed to initial the json
objects as ATM objects right after we read in the file.  So the initial
process is:

 1. Read in json file
 2. Parse input string to JSON
 3. Initialize JSON to array of ATM objects

Also, modified the methods on the controller to return the data
added/updated along with the 'msg' as part of the response.